Json parsing

+++ collection.js
@@ -15,9 +15,11 @@ bay.geom.base.Collection.prototype.getElements = function(){
 
 bay.geom.base.Collection.prototype.clear = function(){
   for(var i=this.list.length - 1;i>=0;i--){
-    this.list[i].deleteElement();
+    if (this.list[i] && this.list[i].deleteElement)
+      this.list[i].deleteElement();
   }
   this.list = [];
+  return this;
 }
 
 bay.geom.base.Collection.prototype.add = function(e){
@@ -55,39 +57,117 @@ bay.geom.base.Collection.prototype.jsonCode = function(){
   return str;
 }
 
+bay.geom.base.Collection.prototype.parseJson = function(str){
+  this.clear();
+  itemList = eval('(' + str + ')');
+  for(var i=0;i<itemList.length;i++){
+    var func = this.getFromJsonFunc(itemList[i].type);
+    if (func)
+      this.list[i] = func(itemList[i], this.list);
+  }
+  return this;
+}
+
 // ********************** add behavior to geometrical elements ************************************* //
 bay.geom.base.PointFree.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "PointFree", "x": ' + this.x + ', "y": ' + this.y + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "PointFree", "x": ' + this.x + ', "y": ' + this.y + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.PointFree.fromJson = function(item, list){
+  var point = new bay.geom.base.PointFree( item.x, item.y);
+  if (item.hidden) point.hidden = true;
+  return point;
 }
 
 bay.geom.base.Point_2l.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "Point_2l", "obj1": ' + list.indexOf(this.obj1) + ', "obj2": ' + list.indexOf(this.obj2) + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "Point_2l", "obj1": ' + list.indexOf(this.obj1) + ', "obj2": ' + list.indexOf(this.obj2) + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.Point_2l.fromJson = function(item, list){
+  var point = new bay.geom.base.Point_2l( list[item.obj1], list[item.obj2]);
+  if (item.hidden) point.hidden = true;
+  return point;
 }
 
 bay.geom.base.Point_2c.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "Point_2c", "obj1": ' + list.indexOf(this.obj1) + ', "obj2": ' + list.indexOf(this.obj2) + ', "num": ' + this.intersectNum + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "Point_2c", "obj1": ' + list.indexOf(this.obj1) + ', "obj2": ' + list.indexOf(this.obj2) + ', "num": ' + this.intersectNum + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.Point_2c.fromJson = function(item, list){
+  var point = new bay.geom.base.Point_2c( list[item.obj1], list[item.obj2], item.num);
+  if (item.hidden) point.hidden = true;
+  return point;
 }
 
 bay.geom.base.Point_lc.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "Point_lc", "obj1": ' + list.indexOf(this.obj1) + ', "obj2": ' + list.indexOf(this.obj2) + ', "num": ' + this.intersectNum + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "Point_lc", "obj1": ' + list.indexOf(this.obj1) + ', "obj2": ' + list.indexOf(this.obj2) + ', "num": ' + this.intersectNum + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.Point_lc.fromJson = function(item, list){
+  var point = new bay.geom.base.Point_lc( list[item.obj1], list[item.obj2], item.num);
+  if (item.hidden) point.hidden = true;
+  return point;
 }
 
 bay.geom.base.LineGeneral.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "LineGeneral", "a": ' + this.a + ', "b": ' + this.b + ', "c": ' + this.c + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "LineGeneral", "a": ' + this.a + ', "b": ' + this.b + ', "c": ' + this.c + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.LineGeneral.fromJson = function(item, list){
+  var line = new bay.geom.base.LineGeneral( item.a, item.b, item.c);
+  if (item.hidden) line.hidden = true;
+  return line;
 }
 
 bay.geom.base.Line_2p.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "Line_2p", "p1": ' + list.indexOf(this.startPoint) + ', "p2": ' + list.indexOf(this.endPoint) + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "Line_2p", "p1": ' + list.indexOf(this.startPoint) + ', "p2": ' + list.indexOf(this.endPoint) + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.Line_2p.fromJson = function(item, list){
+  var line = new bay.geom.base.Line_2p( list[item.p1], list[item.p2]);
+  if (item.hidden) line.hidden = true;
+  return line;
 }
 
 bay.geom.base.Segment.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "Segment", "p1": ' + list.indexOf(this.startPoint) + ', "p2": ' + list.indexOf(this.endPoint) + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "Segment", "p1": ' + list.indexOf(this.startPoint) + ', "p2": ' + list.indexOf(this.endPoint) + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.Segment.fromJson = function(item, list){
+  var line = new bay.geom.base.Segment( list[item.p1], list[item.p2]);
+  if (item.hidden) line.hidden = true;
+  return line;
 }
 
 bay.geom.base.CircleGeneral.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "CircleGeneral", "a": ' + this.a + ', "b": ' + this.b + ', "c": ' + this.c + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "CircleGeneral", "a": ' + this.a + ', "b": ' + this.b + ', "c": ' + this.c + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.CircleGeneral.fromJson = function(item, list){
+  var circle = new bay.geom.base.CircleGeneral( item.a, item.b, item.c);
+  if (item.hidden) circle.hidden = true;
+  return circle;
 }
 
 bay.geom.base.Circle_3p.prototype.toJson = function(list, id){
-  return '{"id": ' + id + ', type": "Circle_3p", "p1": ' + list.indexOf(this.centerPoint) + ', "p2": ' + list.indexOf(this.startPoint) + ', "p3": ' + list.indexOf(this.endPoint) + (this.hidden?', "hidden": true':'') +'}';
+  return '{"id": ' + id + ', "type": "Circle_3p", "p1": ' + list.indexOf(this.centerPoint) + ', "p2": ' + list.indexOf(this.startPoint) + ', "p3": ' + list.indexOf(this.endPoint) + (this.hidden?', "hidden": true':'') +'}';
+}
+
+bay.geom.base.Circle_3p.fromJson = function(item, list){
+  var circle = new bay.geom.base.Circle_3p( list[item.p1], list[item.p2], list[item.p3]);
+  if (item.hidden) circle.hidden = true;
+  return circle;
+}
+
+// ************************** determine parse function ***********************************
+bay.geom.base.Collection.prototype.getFromJsonFunc = function(objName){
+  if (objName=="PointFree") return bay.geom.base.PointFree.fromJson;
+  else if (objName=="Point_2l") return bay.geom.base.Point_2l.fromJson;
+  else if (objName=="Point_2c") return bay.geom.base.Point_2c.fromJson;
+  else if (objName=="Point_lc") return bay.geom.base.Point_lc.fromJson;
+  else if (objName=="LineGeneral") return bay.geom.base.LineGeneral.fromJson;
+  else if (objName=="Line_2p") return bay.geom.base.Line_2p.fromJson;
+  else if (objName=="Segment") return bay.geom.base.Segment.fromJson;
+  else if (objName=="CircleGeneral") return bay.geom.base.CircleGeneral.fromJson;
+  else if (objName=="Circle_3p") return bay.geom.base.Circle_3p.fromJson;
 }
